requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.7
dlib==19.24.6
dnspython==2.7.0
email-validator==2.0.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
humanize==4.12.2
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.9
MarkupSafe==3.0.2
numpy==1.26.0
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.21
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==2.3.7
